refactor: normalize policy-layer timestamps to uint48
Aligns the example policy contracts (RecurringAllowance, SessionPolicy,
PolicyManager) with the canonical Keystore convention that every Unix-second
timestamp is uint48 (period, start, end, validAfter, validUntil, the ONE_TIME
and end sentinels).

- No extra SLOAD/SSTORE: the only stored time struct, RecurringAllowance.PeriodUsage,
  packs as uint48 + uint48 + uint160 = 32 bytes = one slot (was 30 bytes / one slot).
  Limit/Config/TokenLimit/PolicyBinding are memory/calldata; PolicyManager is stateless.
- No digest change: policyConfig and the binding commitment are built with abi.encode,
  which pads every integer to 32 bytes by value, so widening the field type leaves the
  encoded bytes (and thus keccak256) unchanged.
- Also fixes leftover uint40 casts of the uint48 `unlocksAt` in the Keystore tests.

+++ src/policies/PolicyManager.sol
@@ -77,9 +77,9 @@ contract PolicyManager is ReentrancyGuard {
         /// @dev Committed, opaque policy configuration interpreted by `policy`.
         bytes policyConfig;
         /// @dev Earliest timestamp (seconds) at which execution is allowed. Zero = no lower bound.
-        uint40 validAfter;
+        uint48 validAfter;
         /// @dev Timestamp (seconds) at/after which execution is disallowed. Zero = no upper bound.
-        uint40 validUntil;
+        uint48 validUntil;
         /// @dev Salt allowing multiple distinct bindings for the same (account, policy, config).
         uint256 salt;
     }
@@ -90,7 +90,7 @@ contract PolicyManager is ReentrancyGuard {
     event ExecutionSkipped(address indexed account, address indexed policy, bytes32 indexed actorId);
 
     /// @notice The current time is outside the binding's `[validAfter, validUntil)` execution window.
-    error OutsideValidityWindow(uint40 validAfter, uint40 validUntil, uint256 timestamp);
+    error OutsideValidityWindow(uint48 validAfter, uint48 validUntil, uint256 timestamp);
     /// @notice The supplied binding does not recompute to the actor's live signed commitment.
     error BindingCommitmentMismatch(bytes32 expected, bytes32 actual);
     /// @notice {execute} requires `binding.account == msg.sender` (the protocol-dispatched account).
